refactor: 내 퀴즈 요약 기능 폴더 구조 변경
